cleaning up mapview to support marker
