+++ src/containers/MarkerPopup/index.js
@@ -1,0 +1,35 @@
+import React, { Component } from 'react';
+
+class MarkerPopup extends Component {
+  constructor(props){
+    super(props);
+
+      this.state = {
+        data: 'hi'
+      }  
+  }
+
+
+
+
+
+
+
+
+
+  render() {
+    return (
+      <div>
+
+        
+        
+
+
+
+      </div>
+    );
+  }
+}
+
+
+export default MarkerPopup;
